historify/requirements.txt
anyio==4.9.0
APScheduler==3.11.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
llvmlite==0.44.0
MarkupSafe==3.0.2
numba==0.61.2
numpy==2.2.6
openalgo==1.0.15
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
websocket-client==1.8.0
Werkzeug==3.1.3